Use async requests
The non-blocking and async requests prevent freezing the application.
As result of that the methods do not return any return values directly,
so there are callback functions needed to receive data or even any kind
of response.

To provide a fairly generic design the last method argument always has
to be an object defining the callback functions.
This way the param list needs only one param for passing all callback
functions.

In generall the callback functions are the same as jQuery's ajax
API provides, namely *success*, *error* and *complete*.
The callback functions have to provide the same params as jQuery's
pendants.
For more information see: http://api.jquery.com/jquery.ajax/